// src/app/models/current-user.ts
export class CurrentUser {

    userId: string;
    firstName: string;
    lastName: string;
    userName: string;
    email: string;
    active: boolean;
    role: string;
    organizationId: string;
    organizationName: string;
    netLogicPassword: string;
    accepted: boolean;
    phone: string;

    address: string;
    city: string;
    country: string;

    numberOfCoreAgents: number;
    numberOfEnterpriseAgents: number;
    numberOfRecorderAgents: number;
    numberOfOutboundAgents: number;
    numberOfMonths: number;
    startDate: string;
    endDate: string;
    paymentStatus: string;
    remainingDays: number;

    totalCost: string;
}