// src/app/models/payment.ts
export class Payment {

    numberOfCoreAgents: number;
    numberOfEnterpriseAgents: number;
    numberOfRecorderAgents: number;
    numberOfOutboundAgents: number;
    supersCount: number;
    numberOfMonths: number;
    interval: string;
    totalCost: number;
    startDate: string;
    endDate: string;
    freeTrial: boolean;
}